checksScheduleExact test
- test checksScheduleExact
