+++ libtelemetry/src/test/java/com/mapbox/android/telemetry/AlarmReceiverTest.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 
 import org.junit.Test;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -23,5 +24,16 @@ public class AlarmReceiverTest {
     theAlarmReceiver.onReceive(mockedContext, mockedIntent);
 
     verify(mockedSchedulerCallback, times(1)).onPeriodRaised();
+  }
+
+  @Test
+  public void checksSupplyIntent() throws Exception {
+    SchedulerCallback mockedSchedulerCallback = mock(SchedulerCallback.class);
+    AlarmReceiver theAlarmReceiver = new AlarmReceiver(mockedSchedulerCallback);
+
+    Intent actualAlarmIntent = theAlarmReceiver.supplyIntent();
+    Intent expectedAlarmIntent = new Intent("com.mapbox.scheduler_flusher");
+
+    assertEquals(expectedAlarmIntent.getAction(), actualAlarmIntent.getAction());
   }
 }
